relocate getWordValidityGrids
